Split internal/api/auth.go into sessions, middleware, and handlers
Pure file moves within package api: SessionStore and session types move
to sessions.go, auth middleware and path exemptions to middleware.go,
auth.go keeps the HTTP handlers and password helpers. No API changes.
