+++ internal/api/auth.go
@@ -1,336 +1,19 @@
 package api
 
 import (
-	"context"
-	"crypto/rand"
 	"crypto/sha256"
-	"crypto/subtle"
 	"encoding/hex"
 	"encoding/json"
-	"fmt"
 	"log/slog"
 	"net/http"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/JeremiahM37/librarr/internal/config"
 	"github.com/JeremiahM37/librarr/internal/db"
 	"golang.org/x/crypto/bcrypt"
 )
-
-// contextKey is an unexported type for context keys in this package.
-type contextKey string
-
-const (
-	ctxUserID   contextKey = "userID"
-	ctxUserRole contextKey = "userRole"
-	ctxUsername contextKey = "username"
-)
-
-// SessionData holds session metadata.
-type SessionData struct {
-	UserID   int64
-	Username string
-	Role     string
-	Expiry   time.Time
-}
-
-// PendingTOTP holds a pending TOTP verification.
-type PendingTOTP struct {
-	UserID int64
-	Expiry time.Time
-}
-
-// SessionStore manages session-based authentication with user tracking.
-type SessionStore struct {
-	mu          sync.RWMutex
-	sessions    map[string]*SessionData
-	pendingTOTP map[string]*PendingTOTP
-}
-
-// NewSessionStore creates a new session store.
-func NewSessionStore() *SessionStore {
-	s := &SessionStore{
-		sessions:    make(map[string]*SessionData),
-		pendingTOTP: make(map[string]*PendingTOTP),
-	}
-
-	// Periodically clean up expired sessions and pending TOTP tokens.
-	go func() {
-		ticker := time.NewTicker(10 * time.Minute)
-		for range ticker.C {
-			now := time.Now()
-			s.mu.Lock()
-			for token, data := range s.sessions {
-				if now.After(data.Expiry) {
-					delete(s.sessions, token)
-				}
-			}
-			for token, pending := range s.pendingTOTP {
-				if now.After(pending.Expiry) {
-					delete(s.pendingTOTP, token)
-				}
-			}
-			s.mu.Unlock()
-		}
-	}()
-
-	return s
-}
-
-// Create generates a new session token for a user, valid for 24 hours. It
-// returns an error if the system CSPRNG fails, so callers fail closed rather
-// than mint a predictable all-zero token.
-func (s *SessionStore) Create(userID int64, username, role string) (string, error) {
-	b := make([]byte, 32)
-	if _, err := rand.Read(b); err != nil {
-		return "", fmt.Errorf("generate session token: %w", err)
-	}
-	token := hex.EncodeToString(b)
-
-	s.mu.Lock()
-	s.sessions[token] = &SessionData{
-		UserID:   userID,
-		Username: username,
-		Role:     role,
-		Expiry:   time.Now().Add(24 * time.Hour),
-	}
-	s.mu.Unlock()
-
-	return token, nil
-}
-
-// CreatePendingTOTP creates a temporary token for TOTP verification (5 min
-// expiry). Like Create, it fails closed if the CSPRNG errors.
-func (s *SessionStore) CreatePendingTOTP(userID int64) (string, error) {
-	b := make([]byte, 32)
-	if _, err := rand.Read(b); err != nil {
-		return "", fmt.Errorf("generate pending TOTP token: %w", err)
-	}
-	token := hex.EncodeToString(b)
-
-	s.mu.Lock()
-	s.pendingTOTP[token] = &PendingTOTP{
-		UserID: userID,
-		Expiry: time.Now().Add(5 * time.Minute),
-	}
-	s.mu.Unlock()
-
-	return token, nil
-}
-
-// ValidatePendingTOTP checks and consumes a pending TOTP token.
-func (s *SessionStore) ValidatePendingTOTP(token string) (int64, bool) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	pending, ok := s.pendingTOTP[token]
-	if !ok {
-		return 0, false
-	}
-	delete(s.pendingTOTP, token)
-
-	if time.Now().After(pending.Expiry) {
-		return 0, false
-	}
-	return pending.UserID, true
-}
-
-// Get retrieves session data if the token is valid.
-func (s *SessionStore) Get(token string) (*SessionData, bool) {
-	s.mu.RLock()
-	data, ok := s.sessions[token]
-	s.mu.RUnlock()
-
-	if !ok {
-		return nil, false
-	}
-	if time.Now().After(data.Expiry) {
-		s.mu.Lock()
-		delete(s.sessions, token)
-		s.mu.Unlock()
-		return nil, false
-	}
-	return data, true
-}
-
-// Valid checks if a session token is valid and not expired (backward compat).
-func (s *SessionStore) Valid(token string) bool {
-	_, ok := s.Get(token)
-	return ok
-}
-
-// Delete removes a session.
-func (s *SessionStore) Delete(token string) {
-	s.mu.Lock()
-	delete(s.sessions, token)
-	s.mu.Unlock()
-}
-
-// exemptPaths are paths that do not require authentication.
-var exemptPaths = map[string]bool{
-	"/":                 true, // Web UI (handles its own login)
-	"/health":           true,
-	"/api/health":       true,
-	"/api/login":        true,
-	"/api/login/totp":   true,
-	"/api/register":     true,
-	"/api/auth/status":  true,
-	"/readyz":           true,
-	"/api/openapi.json": true, // public API schema for AI/tooling discovery
-}
-
-// isExempt returns true if the path does not require auth.
-func isExempt(path string) bool {
-	if exemptPaths[path] {
-		return true
-	}
-	// Torznab has its own apikey auth (checked inside the handler itself).
-	// Both the canonical path and the Prowlarr-compat /api alias are exempt;
-	// the alias is mounted as exact path /api only, so this does NOT match
-	// /api/search, /api/library, etc.
-	if path == "/api" || strings.HasPrefix(path, "/torznab/") {
-		return true
-	}
-	// Static assets.
-	if strings.HasPrefix(path, "/static/") {
-		return true
-	}
-	// OPDS feeds (e-readers handle auth separately).
-	if strings.HasPrefix(path, "/opds") {
-		return true
-	}
-	// Prometheus metrics.
-	if path == "/metrics" {
-		return true
-	}
-	// OIDC auth endpoints.
-	if strings.HasPrefix(path, "/auth/oidc/") {
-		return true
-	}
-	return false
-}
-
-// authMiddleware returns an HTTP middleware that enforces authentication.
-func authMiddleware(cfg *config.Config, database *db.DB, sessions *SessionStore, next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Check if multi-user is active (any users in DB).
-		userCount, _ := database.CountUsers()
-		multiUser := userCount > 0
-
-		// Trusted reverse-proxy SSO headers should short-circuit the normal
-		// login flow when OIDC is configured. This lets Authentik-backed
-		// deployments log users in transparently instead of requiring a second
-		// click on the Librarr login button.
-		if cfg != nil && cfg.HasOIDCProxyHeaders() {
-			username := proxyIdentityFromRequest(r)
-			if username != "" {
-				if user, err := resolveOIDCUser(cfg, database, username); err == nil && user != nil {
-					if sessions != nil {
-						if ensureSessionForUser(w, r, sessions, user) {
-							_ = database.UpdateLastLogin(user.ID)
-						}
-					}
-					ctx := context.WithValue(r.Context(), ctxUserID, user.ID)
-					ctx = context.WithValue(ctx, ctxUserRole, user.Role)
-					ctx = context.WithValue(ctx, ctxUsername, user.Username)
-					next.ServeHTTP(w, r.WithContext(ctx))
-					return
-				} else if err != nil && cfg != nil && cfg.HasOIDC() {
-					slog.Warn("proxy SSO login rejected", "username", username, "error", err)
-				}
-			}
-		}
-
-		// No multi-user, no legacy auth, no API key: the instance is open.
-		// Treat the local caller as an admin rather than passing through
-		// role-less, so admin-gated routes (e.g. POST /api/settings) work on
-		// userless instances instead of failing requireAdmin with a 403.
-		if !multiUser && !cfg.HasAuth() && !cfg.HasAPIKey() {
-			ctx := context.WithValue(r.Context(), ctxUserRole, "admin")
-			ctx = context.WithValue(ctx, ctxUsername, "local")
-			next.ServeHTTP(w, r.WithContext(ctx))
-			return
-		}
-
-		// Exempt paths always pass through.
-		if isExempt(r.URL.Path) {
-			next.ServeHTTP(w, r)
-			return
-		}
-
-		// Check API key (header or query param) -- machine-to-machine auth.
-		if cfg.HasAPIKey() {
-			apiKey := r.Header.Get("X-Api-Key")
-			if apiKey == "" {
-				apiKey = r.URL.Query().Get("apikey")
-			}
-			if subtle.ConstantTimeCompare([]byte(apiKey), []byte(cfg.APIKey)) == 1 {
-				// API key users get admin-level access.
-				ctx := context.WithValue(r.Context(), ctxUserRole, "admin")
-				ctx = context.WithValue(ctx, ctxUsername, "api")
-				next.ServeHTTP(w, r.WithContext(ctx))
-				return
-			}
-		}
-
-		// Check session cookie for multi-user mode.
-		if multiUser {
-			cookie, err := r.Cookie("librarr_session")
-			if err == nil {
-				if data, ok := sessions.Get(cookie.Value); ok {
-					ctx := context.WithValue(r.Context(), ctxUserID, data.UserID)
-					ctx = context.WithValue(ctx, ctxUserRole, data.Role)
-					ctx = context.WithValue(ctx, ctxUsername, data.Username)
-					next.ServeHTTP(w, r.WithContext(ctx))
-					return
-				}
-			}
-		}
-
-		// Legacy single-user session auth (when no multi-user DB users exist).
-		if !multiUser && cfg.HasAuth() {
-			cookie, err := r.Cookie("librarr_session")
-			if err == nil && sessions.Valid(cookie.Value) {
-				ctx := context.WithValue(r.Context(), ctxUserRole, "admin")
-				ctx = context.WithValue(ctx, ctxUsername, cfg.AuthUsername)
-				next.ServeHTTP(w, r.WithContext(ctx))
-				return
-			}
-		}
-
-		// No valid auth found.
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusUnauthorized)
-		json.NewEncoder(w).Encode(map[string]interface{}{
-			"success": false,
-			"error":   "Authentication required",
-		})
-	})
-}
-
-// requireAdmin is middleware that checks if the current user has admin role.
-func requireAdmin(next http.HandlerFunc) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		role, _ := r.Context().Value(ctxUserRole).(string)
-		if role != "admin" {
-			writeJSON(w, http.StatusForbidden, map[string]interface{}{
-				"success": false,
-				"error":   "Admin access required",
-			})
-			return
-		}
-		next(w, r)
-	}
-}
-
-// getUserIDFromContext extracts the user ID from the request context.
-func getUserIDFromContext(r *http.Request) int64 {
-	id, _ := r.Context().Value(ctxUserID).(int64)
-	return id
-}
 
 // hashPassword hashes a password using bcrypt.
 func hashPassword(password string) (string, error) {
